requirements.txt
anyio==3.6.2
appdirs==1.4.4
beautifulsoup4==4.12.2
black==23.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cryptography==40.0.2
fastapi==0.95.1
flake8==6.0.0
frozendict==2.3.8
h11==0.14.0
html5lib==1.1
httpcore==0.17.0
httpx==0.24.0
idna==3.4
isort==5.12.0
lxml==4.9.2
mccabe==0.7.0
multitasking==0.0.11
mypy==1.2.0
mypy-extensions==1.0.0
numpy==1.24.3
packaging==23.1
pandas==2.0.1
pathspec==0.11.1
platformdirs==3.5.0
pycodestyle==2.10.0
pycparser==2.21
pydantic==1.10.7
pyflakes==3.0.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
requests==2.29.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
starlette==0.26.1
tomli==2.0.1
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
uvicorn==0.22.0
webencodings==0.5.1
yfinance==0.2.18
